working split_meta
